# rafthttp: reduce allocs in msgappv2