The patch decreases the allocs when sending one AppEntry in msgappv2
stream from 30 to 9. This helps reduce CPU load when etcd is under
high write load.

before:
```
BenchmarkSendingMsgApp	 100000	    18333 ns/op	  3.49 MB/s	   1497 B/op	     30 allocs/op
```

after:
```
BenchmarkSendingMsgApp	  100000	     16492 ns/op	   3.88 MB/s	    1040 B/op	       9 allocs/op
```

/cc @barakmich