@barakmich 2015-05-12
Huh. I'll take a 10% throughput bump any day. LGTM